M06: six more functions byte-matched, 21 of 55 (issue #84)
Lands four further carves, each fnmatch-verified before carving and
byte-identical after `make clean && make compare` (knidl.gba: OK):

- src/terrain_21130.c  0x08021130-0x0802136C: sub_08021130, the player's
  room probe (sub_0802136c's body with the tile-set 64..79 / 192..207
  offset lookup in front and the gUnk_08733AF0 attribute copy behind).
- src/terrain_1c30c.c  0x0801C30C-0x0801C444: sub_0801c30c and
  sub_0801c3a4, the per-frame entry points (box load, room-relative
  position from Task.unk4C/unk50 and unk54/unk58, probes, write-back).
- src/terrain_1c51c.c  0x0801C51C-0x0801C690: sub_0801c51c (box offsets
  and task fields into the room-descriptor cells, result block cleared)
  and sub_0801c5c8 (results back into the task, mirror into gUnk_03005550).
- src/terrain_1c8dc.c  0x0801C8DC-0x0801C930: sub_0801c8dc.

The five 32-byte attribute helpers in terrain_214e0.c now return s32 with
an s8 element type instead of s8 with a u8 element type: same bytes, but
the callers (sub_0801c690 and friends) compare the raw r0, which only an
int prototype reproduces (lesson 3.289). terrain_214e0.c re-verified.

Progress: 492688 of 857388 code bytes in src (57.46%). Seven more IWRAM
cells / ROM tables registered in split_config.json data_symbols.

Lessons 3.291 (a zero variable's position among the neighbouring stores
picks its hi-register reload temp) and 3.292 (a stored value that is
reused is a chained assignment, kept s32). rom-map M06 entry and the
module-map row updated to 21/55.

Still in progress locally: sub_0802136c (8 differing bytes: the mask
constant's materialisation order vs the r8/r9 tie), sub_0801c444 (6
bytes: operand order of four narrowed adds), sub_0801c690/sub_0801c7cc
(the k & attribute test and a cross-jumped tail), sub_080207a0
(block layout).

+++ src/terrain_214e0.c
@@ -14,7 +14,8 @@
  *
  * Matching notes (agbcc -O2 -mthumb-interwork -fprologue-bugfix):
  *  - the helpers return int, not s8: callers compare the result without
- *    re-extending it (sub_0802069c);
+ *    re-extending it (sub_0802069c, sub_0801c690); the sign extension comes
+ *    from the s8 element type (lesson 3.289);
  *  - `p = table[i]; return p[j];` orders the pointer load before the index
  *    load; writing table[i][j] in one expression loads the index first;
  *  - the row/column cell access is `(&gUnk_03005660[idx])[x]` with
@@ -26,11 +27,11 @@
 
 
 /* ROM pointer tables: one entry per tile set, each pointing at a byte table. */
-extern u8 *const gUnk_08734BF0[];
-extern u8 *const gUnk_08733BF0[];
-extern u8 *const gUnk_08733FF0[];
-extern u8 *const gUnk_087343F0[];
-extern u8 *const gUnk_087347F0[];
+extern s8 *const gUnk_08734BF0[];
+extern s8 *const gUnk_08733BF0[];
+extern s8 *const gUnk_08733FF0[];
+extern s8 *const gUnk_087343F0[];
+extern s8 *const gUnk_087347F0[];
 extern s8 *const gUnk_087330F0[];
 extern s8 *const gUnk_087328F0[];
 
@@ -241,33 +242,33 @@ s32 sub_080218f8(u32 x, u32 y)
     return gUnk_087330F0[tile][off];
 }
 
-s8 sub_08021970(u16 a)
+s32 sub_08021970(u16 a)
 {
-    u8 *p = gUnk_08734BF0[a];
+    s8 *p = gUnk_08734BF0[a];
     return p[gUnk_03005508];
 }
 
-s8 sub_08021990(u16 a)
+s32 sub_08021990(u16 a)
 {
-    u8 *p = gUnk_08733BF0[a];
+    s8 *p = gUnk_08733BF0[a];
     return p[gUnk_03005508];
 }
 
-s8 sub_080219b0(u16 a)
+s32 sub_080219b0(u16 a)
 {
-    u8 *p = gUnk_08733FF0[a];
+    s8 *p = gUnk_08733FF0[a];
     return p[gUnk_03005508];
 }
 
-s8 sub_080219d0(u16 a)
+s32 sub_080219d0(u16 a)
 {
-    u8 *p = gUnk_087343F0[a];
+    s8 *p = gUnk_087343F0[a];
     return p[gUnk_03005508];
 }
 
-s8 sub_080219f0(u16 a)
+s32 sub_080219f0(u16 a)
 {
-    u8 *p = gUnk_087347F0[a];
+    s8 *p = gUnk_087347F0[a];
     return p[gUnk_03005508];
 }
